Fix function calls.
Need to use `LLVMGlobalGetValueType` instead of `LLVMTypeOf` for getting
the type of a function, since functions are always `GlobalValue` which
are of type `ptr` (which is opaque). Using `GlobalValue::getValueType()`
fixes the issue.

A simple `cloak` program like below is able to successfully compile with
this change:

```
fn foo(): i32 {
    return 1;
}

fn main(): i32 {
    return foo();
}
```